--- src/pager/add_header_guard.txt
@@ -1,1 +1,0 @@
-Eyes up guardian
