+++ src/pager/db/catalog/README.md
@@ -1,4 +1,4 @@
 Catalog pages are special pages implemented as real B+ Tree backed Index tables.
 
-The only difference is that their initialization is hardcoded into the database engine - such that whenever it boots it calls all the 3 initialization methods to immediately populate the database with these values.
+The only difference is that their initialization is hardcoded into the database engine - such that whenever it boots it calls all the 3 initialization methods to immediately populate the database with indexes based on their primary `id` keys.
