+++ src/pager/db/data/page_format.h
@@ -4,7 +4,7 @@
  */
 
 #define MAX_DATA_SLOTS 256  // Max number of slots per data page
-#include "psql_types.h"  // For the base types, NULL, PTR, TEXT, INT
+#include "types/psql_types.h"  // For the base types, NULL, PTR, TEXT, INT
 
 typedef struct {
     uint16_t num_slots;
