+++ src/pager/pager.c
@@ -9,43 +9,6 @@
 #include <errno.h>
 #include "db/catalog/catalog.h"
 
-/* CRC32 table for checksum calculation */
-static uint32_t crc32_table[256];
-static int crc32_initialized = 0;
-
-/* Initialize CRC32 table */
-static void init_crc32_table() {
-    if (crc32_initialized) return;
-    
-    uint32_t polynomial = 0xEDB88320;
-    for (uint32_t i = 0; i < 256; i++) {
-        uint32_t c = i;
-        for (size_t j = 0; j < 8; j++) {
-            if (c & 1) {
-                c = polynomial ^ (c >> 1);
-            } else {
-                c >>= 1;
-            }
-        }
-        crc32_table[i] = c;
-    }
-    crc32_initialized = 1;
-}
-
-/* Calculate CRC32 checksum */
-uint32_t calculate_crc32(const void* data, size_t length) {
-    if (!crc32_initialized) init_crc32_table();
-    
-    const uint8_t* buf = (const uint8_t*)data;
-    uint32_t crc = 0xFFFFFFFF;
-    
-    for (size_t i = 0; i < length; i++) {
-        crc = crc32_table[(crc ^ buf[i]) & 0xFF] ^ (crc >> 8);
-    }
-    
-    return ~crc;
-}
-
 void init_free_page_map(Pager* pager) {
     pager->free_page_map = radix_tree_create();
     
@@ -585,6 +548,10 @@ PSqlStatus pager_init_new_db(Pager* pager) {
     return pager_sync(pager);
 }
 
+
+// Perform checks on the validity of a PreSeQl DB file
+// These are all on the magic number (invalid file type)
+// CRC32 on header (except the header checksum - we zero before recalculating)
 PSqlStatus pager_verify_db(Pager* pager) {
     if (!pager) return PSQL_ERROR;
     
@@ -607,7 +574,7 @@ PSqlStatus pager_verify_db(Pager* pager) {
     return PSQL_OK;
 }
 
-/* Implementation of psql_open and psql_close */
+/* Implementation of psql_open and psql_close from preseql.h */
 PSqlStatus psql_open(PSql* db) {
     if (!db || !db->filename) return PSQL_ERROR;
